Update PyQSOFit to find the luminosity at 2500A.
In SDSS DR7 most sources don't extend their optical spectra up to 3000 mostly for high redshift sources and the latest versions where L1350 is also not available. Hence the luminosity parameter obtained from this could not be used to generate meaningful results. Including a min wavelength luminosity such as 2500A,will help to improve better usage of the luminosity parameter as the spectral slope. Furthermore, the suggestion includes making PolyFit as False in default as this is impacting a great change than the actual spectral slope value.