Bugfix
1) Add close fig function to avoid memory leak.
When savefig is set to True, the figure will be closed automatically. If one would like to show the figure first, he/she can set savefig to False, call self.fig to exhibit the figure and save it manually.
2) Change the package sfdmap to sfdmap2.
Since the sfdmap is no longer maintained and has been conflicted with the latest numpy package, we change it to a forked repository sfdmap2. ref: https://github.com/AmpelAstro/sfdmap2
3) Set the default jitter to 0.0 to avoid the unrepeatable fitting results. (temporary)
We do think the jitter will help to get unbiased fitting results especially when most of the initial parameters are 0. However, since the jitter is not a physical parameter, it is not reasonable to set it as a default global absolute value. Besides, variable results in repeated running are confusing to users. We will find out a more feasible way to use the jitter in the future.